Added Deepsleep for ESP Boards,
Moved LoRaWAN Keys form main.cpp to config.h, this will make it
easier, to handle multiple boards with diffrent keys.

No we porvide some preprocessor conditions for diffrent boards,
defined in config.h.
We use Adafruit_SleepyDog library when FEATHERWING is defined.
We use esp_light_sleep_start when TTGOLORA is defined.

Changed messages in getSensorType();

Fixed a bug in getSensorValues(); Now we check for the sensor type
and init the sensor, after powering it up agian. This makes sure we
get the right values from the sensor.